requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
DateTime==5.5
debugpy==1.8.13
decorator==5.2.1
executing==2.2.0
ExifRead==3.0.0
ffmpeg-python==0.2.0
fonttools==4.56.0
future==1.0.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
piexif==1.1.3
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
pyzmq==26.3.0
seaborn==0.13.2
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tzdata==2025.2
wcwidth==0.2.13
zope.interface==7.2
